+++ internal/model/order_item.go
@@ -7,11 +7,11 @@ import (
 type OrderItem struct {
 	gorm.Model
 
-	OrderID   uint64 `gorm:"not null;index;column:order_id"`
-	ProductID uint64 `gorm:"not null;index;column:product_id"`
+	OrderID   uint `gorm:"not null;index;column:order_id"`
+	ProductID uint `gorm:"not null;index;column:product_id"`
 
 	ProductName string `gorm:"type:varchar(100);not null;column:product_name"`
-	UnitPrice   int64  `gorm:"not null;column:unit_price"`
-	Quantity    int    `gorm:"not null;column:quantity"`
-	TotalPrice  int64  `gorm:"not null;column:total_price"`
+	UnitPrice   uint64 `gorm:"not null;column:unit_price"`
+	Quantity    uint   `gorm:"not null;column:quantity"`
+	TotalPrice  uint64 `gorm:"not null;column:total_price"`
 }
